Added a lot of fluid simulation
Added a parametric study that tracks required input torque in relation to impeller velocity. It also tracks the output volume rate. We can use this setup to optimize the physical model around the power curve of the motor. I also included an excel sheet that shows the data generated by the fluid sim